Refactor out a generic HTTPS request function
To facilitate local debugging, let's factor out as much code into
modules as possible, starting with the functionality to send an HTTPS
request (which is quite low-level, albeit less so than
`https.request()`).

This commit is best viewed with `--color-moved`.
