+++ LeaveMeAloneGitHubApp/https-request.js
@@ -1,0 +1,44 @@
+module.exports = async (context, requestMethod, requestPath, body, headers) => {
+    return new Promise((resolve, reject) => {
+        const https = require('https')
+        const request = https.request({
+            host: "api.github.com",
+            port: 443,
+            path: requestPath,
+            method: requestMethod || "GET",
+            headers: {
+                "User-Agent": "curl/7.68.0",
+                Accept: "application/vnd.github+json",
+		...headers
+            }
+        }, (res, e) => {
+            if (e) {
+                reject(e)
+                return
+            }
+            context.log(`${requestPath} returned ${res.statusCode}`)
+            context.log(res.headers)
+            res.setEncoding('utf8')
+            var response = ''
+            res.on('data', (chunk) => {
+                response += chunk
+            })
+            res.on('end', () => {
+                if (!response) {
+                    resolve(response)
+                    return
+                }
+                try {
+                    resolve(JSON.parse(response))
+                } catch (e) {
+                    reject(e)
+                }
+            })
+            res.on('error', (e) => {
+                reject(e)
+            })
+        })
+        if (body) request.write(body)
+        request.end()
+    })
+}
